Move override application logic to configuration class
While the normal expected use case for an override is to apply it to a request, there are uses cases where we might want to apply an override to a non-request context. For example, we have a script that runs outside of the main rails app to generate some static configs we copy/paste into some static files.

It would look something like this:

```
_, static_csp_policy = SecureHeaders::CSP.make_header(SecureHeaders::Configuration.get.apply_overrides(:static_file_policy).csp, nil)
...
%q(<meta http-equiv="Content-Security-Policy" content="#{static_csp+policy"})
```