Hide default config behind a private class method
We don't expect public callers to actually access the default config directly.
So, we hide it behind a private class method. There is only one caller in the
entire codebase that is not a test. It does make tests a bit uglier, but is
probably worth it for the benefits of hiding it from the public API.