Deadlink in readme file in this repo to https://github.com/sourceclear/headlines which does not exist - Status code [404:NotFound]
In this repo's readme there is a link to: https://github.com/sourceclear/headlines
This is in section for "Similar libraries".
However that is a deadlink and gives Status code [404:NotFound].

Looking at https://github.com/sourceclear there does not appear to be any similar named repos belonging to them.
So I assume this is not a simple typo and that this repo has been removed.

Perhaps the simplest thing to do here is to just remove this "Similar libraries" reference all together?
So that is what I have done in this PR.

However please feel free to suggest an alternate approach.

## extra

I recently created a tool that makes use of the GitHub api to find repos that match certain parameters. Then this tool checks the repo's readme files for dead links: 
http://githubreadmechecker.com/Home/Search?SingleRepoUri=https%3a%2f%2fgithub.com%2fgithub%2fsecure_headers

It can be used to re-check this repo via this url: http://githubreadmechecker.com/Home/Search?SingleRepoUri=https%3a%2f%2fgithub.com%2fgithub%2fsecure_headers

I'd love to hear some feedback on this project so please feel free to to share some thoughts, or even start a discussion on the repo's main GitHub page.